adapter: share utilities between hybrid and vox
